backend/requirements.txt
flask
flask_pymongo
fastapi
uvicorn
motor
werkzeug
pyjwt
python-dotenv